Use a flatten directory for all adoc files
This commit reworks how asciidoctor files are processed,
by allowing to structure docs using subdirectories, while
maintaining consistent locations for resources, images, samples,
...

This works by flattening all adoc files into a resources
directory before processing. It means that while we can
use a directory tree to organize docs, we must make sure
not to use the same file name twice.

This makes it possible to compose files more easily. For
example, the dependency management docs can be read as
a single file, or individual files, or as chapters.

Technically speaking, we could remove the "userguideSinglePage"
task, but it's kept just because it's significantly slower
to build.